Refactor data quality checks
